pristine-tar: add support for recompressing tarballs
This solves the issue when pristine-tar cannot reproduce a tarball in
any way. Storing a tarball that is not exactly the same as the one
provided by upstream is better than storing nothing at all.

This can also be used to improve delta sizes. By default, deltas larger
than 500kB or 30% of the size of the original tarball will cause the
original tarball to be recompressed and a new delta be generated based
on that recompressed tarball.